requirements.txt
# Audio splitting tool dependencies

# Speech recognition
openai-whisper>=20231117

# Audio processing
pydub>=0.25.1
librosa>=0.10.0

# Numerical computing
numpy>=1.21.0
scipy>=1.7.0

# Deep learning framework (Whisper dependency)
torch>=1.10.0
torchaudio>=0.10.0

# Note: FFmpeg is required for processing MP3 and other audio formats
# Windows users can download from https://ffmpeg.org/download.html
# Or use: conda install ffmpeg
# Or use: choco install ffmpeg (requires Chocolatey)